refactor energy profile graph
