+++ app/frontend/types/components/EnergyProfileGraph.ts
@@ -1,3 +1,4 @@
+import { GraphDataPoint } from "./Graph";
 export interface EnergyProfileGraphProps {
     enabled?: boolean;
 }
@@ -17,4 +18,12 @@ export interface GraphMetadata {
 export interface GraphResponse {
     metaData: GraphMetadata;
     graphData: GraphData[];
+}
+
+export interface EnergyProfileChartProps {
+    chartData: GraphDataPoint[];
+    metadata: GraphMetadata;
+    dataKeys: string[];
+    hasBasislast: boolean;
+    colors: string[];
 }
